Add post reaction handler
Adding a reaction when the reaction icon is
clicked, if the reaction doesn't have a reaction
of the user logged, the reaction is added, but, if
the reaction has a reaction by the user logged,
the reaction is deleted.

The reaction handler only works when is a user is
logged on the page.

  - Adding actions to control reactions
  - Updating posts vuex module to add or delete a reaction
  - Adding the reaction handler in the post details page