Add posts dashboard for the user logged
Dividing the public posts and the private publications
that belong to the registered user. Public publications
are all those that have the "public" and "publication"
labels. Private publications are all those that have the
"public" label and the owner is the username of the
user who is logged in.

Posts actions of the posts vuex module have been separated
in 'public' and 'private' types, exist one list for each
type.

  - Updating the mutations types
  - Updating the posts vuex module
  - Adding the dashboard carpet with the files
    to add, list, and edit the posts that belong
    to the logged in user
  - Updating the private routes
  - Updating the public posts index page to get the
    public posts from the public vuex state