Add pagination Header in index and dashboard posts lists
Header with an input to select how many items will be
displayed on the list. The options are extracted from
the default-data file.

The count of how many items will be showed is added
in the URL with the 'items' query parameter, the way to
work is exactly as the 'page' query parameter.

  - Adding options availabled to select how many items
    will be displayed
  - Adding PaginationHead component that have the select
    with the option and it receives the 'itemsPerPage' prop
  - Adding PaginationHead in dashboard posts list
  - Adding PaginationHead in index public posts list