Update pagination feature in dashboard and index public post pages
Adding changes to get elements from the GitHub API, because
only returns 30 elements for request and not shows all registers,
the solution is to add a query parameter called 'page' that it is the
method to paginate registers in the API.

The response of the request contains in its headers an attribute
'link' of the string type, this attribute has the number of the
last page of the query that has been made, but, the last page only
will be extract when the request doesn't have the param 'page', so
two requests are made, one without 'page' parameter and other with
'page' parameter.

The number of page is used to determinate how many elements need
be created to the pagination component.

On every change of the param 'page' in the URL, a request will be made
to get all elements of that page.

  - Removing default options to paginate and only letf one that is
    to 30 items because that is the limit of elements returned in
    a response request
  - Updating dashboard page to get the elements of the 'page' param,
    and made a request when the query param 'page' and 'tags' change
  - Updating index public list page to get the elements of the 'page'
    param, and made a request when the query param 'page' and 'tags'
    change
  - Updating posts vuex module to hability to made a request without
    vuex state restrictions