Update AddEditPostEditor to get a post from the API
The publication doesn't extracted from the vuex state
anymore, now is taken directly from the API with a request.

Actions have been added in posts vuex module to get a post,
the post is formatted and added to the vuex state, that
state is taken in AddEditPostEditor and public post
details page and formatted inside them.

  - Updating AddEditPostEditor to get a post from the GitHub API
  - Updating post details page to get, process the post and
    displayed to the user, the post only will be showed if
    is open and public
  - Updating AuthorInformation to not show loading component
    when the post not exist or not pass all validations
  - Adding actions in posts vuex module to get a post from the API